requirements.txt
fastapi
uvicorn
anthropic
asyncpg
python-jose[cryptography]
passlib[bcrypt]
reportlab
python-dotenv
